requirements.txt
alabaster==0.7.13
annotated-types==0.5.0
autopep8==2.0.4
Babel==2.14.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
dgl-cu116==0.9.1.post1
dglgo==0.0.2
dgllife==0.3.2
docutils==0.19
future==1.0.0
hyperopt==0.2.7
idna==3.10
imagesize==1.4.1
importlib-metadata==6.7.0
isort==5.11.5
Jinja2==3.1.5
joblib==1.3.2
littleutils==0.2.2
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==9.1.0
networkx==2.6.3
numpy==1.21.6
numpydoc==1.5.0
ogb==1.3.6
outdated==0.2.2
packaging==21.3
pandas==1.3.5
Pillow==9.5.0
psutil==7.0.0
py4j==0.10.9.9
pycodestyle==2.10.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
rdkit==2023.3.2
rdkit-pypi==2022.9.5
requests==2.31.0
rich==13.8.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
scikit-learn==1.0.2
scipy==1.7.3
shellingham==1.5.4
six==1.17.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
